pyproject.toml
[project]
name = "bg-remover"
version = "0.1.0"
description = "Secure local AI tool for batch background removal using rembg and Pillow"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Andrei Cristian", email = "andrei@keibisoft.com" }
]
keywords = ["ai", "background-removal", "rembg", "cli"]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "rembg>=2.0.50",
    "pillow>=10.2.0",
]

[project.optional-dependencies]
cpu = ["onnxruntime>=1.17.0"]
gpu = ["onnxruntime-gpu>=1.17.0"]

[project.scripts]
bg-remover = "bg_remover:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["input_frames*", "frames*"]

[tool.uv]
package = true
